@@ -1,15 +1,6 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
-require "rails"
-# Pick the frameworks you want:
-require "active_model/railtie"
-require "active_job/railtie"
-require "active_record/railtie"
-require "action_controller/railtie"
-require "action_mailer/railtie"
-require "action_view/railtie"
-require "sprockets/railtie"
-# require "rails/test_unit/railtie"
+require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -17,6 +8,10 @@ Bundler.require(*Rails.groups)
 
 module Cforgoodwebsite
   class Application < Rails::Application
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration should go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded.
+
     config.i18n.enforce_available_locales = true
     config.embed_authenticity_token_in_remote_forms = true
 
@@ -54,4 +49,5 @@ module Cforgoodwebsite
     #background jobs
     config.active_job.queue_adapter = :sidekiq
   end
+
 end
