@@ -1,3 +1,3 @@
 # Protect against injection attacks
 # http://www.kb.cert.org/vuls/id/380039
-ActionDispatch::ParamsParser::DEFAULT_PARSERS.delete(Mime::XML)
+# ActionDispatch::ParamsParser::DEFAULT_PARSERS.delete(Mime::XML)
