@@ -14,7 +14,7 @@ IntercomRails.config do |config|
   # This is required for some Intercom rake tasks like importing your users;
   # you can generate one at https://app.intercom.io/apps/api_keys.
   #
-  config.api_key = ENV["INTERCOM_API_KEY"]
+  # config.api_key = ENV["INTERCOM_API_KEY"]
 
   # == Enabled Environments
   # Which environments is auto inclusion of the Javascript enabled for
